Use right yarn path on apply
Currently the apply function got the path to yarn hard-coded even if
there are paths defined within the install package. I know that the
paths differe a little bit as the hard-coded path doesn't include the
.js suffix, but this difference doesn't break any functionality. With
this change somebody who is compiling the tool on his own is able to
manipulate the required paths.

Signed-off-by: Thomas Boerger <thomas@webhippie.de>